(Console) Magic-mapping colors in "feature" option
Allows the magic-mapping color of the "feature" option to have an effect.

For instance, if your rcfile has the line
    feature += rock wall {,+,,cyan}
Then you would expect to get +'s as the glyphs for magically-mapped rock
walls and to get cyan as the color of these magic-mapped walls. However,
the magic-mapping color was normally ignored (due to not being called),
so it defaulted to the seen_colour() -- usually blue or darkgrey.

Console players should now be able to change the color of magic-mapped
features, which might be useful if they've played the tiles versions and
are used to the standout color of magic-mapped features.